Rendering Pipeline actually doesn't crash!
At least it can run through trying to render an RObj without crashing. Actually rendering it is a different story.

Big shader overhauls will be necessary, and also how shaders and vertexarrays interact with eachother. I'm not 100% sure what is leading to nothing being rendered. 

Next up:
- Do the shaders stuff decently, the idea is that shaders and pipelines get created as soon as a certain robj is created that needs them. 
- How to actually allow for rendering to the mime output. Right now it just selects the first target of the first pipeline, maybe it would be nicer it is a welldefined fbo inside the screen or something, like the main color fbo.
- figure out a way to get rid of the data dict, it's making my head spin, very difficult to understand what is going on
