+++ src/glbackend/GLAbstraction/GLShader.jl
@@ -42,62 +42,36 @@ function getinfolog(obj::GLuint)
     end
 end
 
-function iscompiled(shader::GLuint)
-    success = GLint[0]
-    glGetShaderiv(shader, GL_COMPILE_STATUS, success)
-    return first(success) == GL_TRUE
-end
-islinked(program::GLuint) = glGetProgramiv(program, GL_LINK_STATUS) == GL_TRUE
-
-function createshader(shadertype::GLenum)
-    shaderid = glCreateShader(shadertype)
-    @assert shaderid > 0 "opengl context is not active or shader type not accepted. Shadertype: $(GLENUM(shadertype).name)"
-    shaderid::GLuint
-end
-function createprogram()
-    program = glCreateProgram()
-    @assert program > 0 "couldn't create program. Most likely, opengl context is not active"
-    program::GLuint
-end
-
-shadertype(s::Shader) = s.typ
-function shadertype(f::File{format"GLSLShader"})
-    shadertype(file_extension(f))
-end
-function shadertype(ext::AbstractString)
-    ext == ".comp" && return GL_COMPUTE_SHADER
-    ext == ".vert" && return GL_VERTEX_SHADER
-    ext == ".frag" && return GL_FRAGMENT_SHADER
-    ext == ".geom" && return GL_GEOMETRY_SHADER
-    error("$ext not a valid extension for $f")
-end
-
-#Implement File IO interface
-function load(f::File{format"GLSLShader"})
-    fname = filename(f)
-    source = open(readstring, fname)
-    compile_shader(fname, source)
-end
-function save(f::File{format"GLSLShader"}, data::Shader)
-    s = open(f, "w")
-    write(s, data.source)
-    close(s)
-end
-
-function uniformlocations(nametypedict::Dict{Symbol, GLenum}, program)
-    result = Dict{Symbol, Tuple}()
-    texturetarget = -1 # start -1, as texture samplers start at 0
-    for (name, typ) in nametypedict
-        loc = get_uniform_location(program, name)
-        str_name = string(name)
-        if istexturesampler(typ)
-            texturetarget += 1
-            result[name] = (loc, texturetarget)
-        else
-            result[name] = (loc,)
-        end
+struct Shader
+    name::Symbol
+    source::Vector{UInt8}
+    typ::GLenum
+    id::GLuint
+    context::AbstractContext
+    function Shader(name, source, typ, id)
+        new(name, source, typ, id, current_context())
     end
-    return result
+end
+function Shader(name, source::Vector{UInt8}, typ)
+    compile_shader(source, typ, name)
+end
+name(s::Shader) = s.name
+
+import Base: ==
+
+function (==)(a::Shader, b::Shader)
+    a.source == b.source && a.typ == b.typ && a.id == b.id && a.context == b.context
+end
+
+function Base.hash(s::Shader, h::UInt64)
+    hash((s.source, s.typ, s.id, s.context), h)
+end
+
+
+function Base.show(io::IO, shader::Shader)
+    println(io, GLENUM(shader.typ).name, " shader: $(shader.name))")
+    println(io, "source:")
+    print_with_lines(io, String(shader.source))
 end
 
 abstract type AbstractLazyShader end
@@ -111,9 +85,6 @@ struct LazyShader <: AbstractLazyShader
     end
 end
 
-gl_convert(shader::Program, data) = shader
-
-
 
 
 # caching templated shaders is a pain -.-
@@ -123,7 +94,6 @@ gl_convert(shader::Program, data) = shader
 const _template_cache = Dict{String, Vector{String}}()
 # path --> Dict{template_replacements --> Shader)
 const _shader_cache = Dict{String, Dict{Any, Shader}}()
-const _program_cache = Dict{Any, Program}()
 
 
 function empty_shader_cache!()
@@ -149,6 +119,12 @@ function compile_shader(path, source_str::AbstractString)
     source = Vector{UInt8}(source_str)
     name = Symbol(path)
     compile_shader(source, typ, name)
+end
+
+function createshader(shadertype::GLenum)
+    shaderid = glCreateShader(shadertype)
+    @assert shaderid > 0 "opengl context is not active or shader type not accepted. Shadertype: $(GLENUM(shadertype).name)"
+    shaderid::GLuint
 end
 
 function get_shader!(path, template_replacement, view, attributes)
@@ -179,36 +155,6 @@ function get_template!(path, view, attributes)
     end
 end
 
-
-function compile_program(shaders, fragdatalocation)
-    # Remove old shaders
-    program = createprogram()
-    #attach new ones
-    foreach(shaders) do shader
-        glAttachShader(program, shader.id)
-    end
-
-    #Bind frag data
-    for (location, name) in fragdatalocation
-        glBindFragDataLocation(program, location, ascii(name))
-    end
-
-    #link program
-    glLinkProgram(program)
-    if !GLAbstraction.islinked(program)
-        error(
-            "program $program not linked. Error in: \n",
-            join(map(x-> string(x.name), shaders), " or "), "\n", getinfolog(program)
-        )
-    end
-    # Can be deleted, as they will still be linked to Program and released after program gets released
-    #foreach(glDeleteShader, shader_ids)
-    # generate the link locations
-    nametypedict = uniform_name_type(program)
-    uniformlocationdict = uniformlocations(nametypedict, program)
-    Program(program, shaders, nametypedict, uniformlocationdict)
-end
-
 function get_view(kw_dict)
     _view = kw_dict[:view]
     extension = is_apple() ? "" : "#extension GL_ARB_draw_instanced : enable\n"
@@ -217,12 +163,13 @@ function get_view(kw_dict)
     _view
 end
 
+#TODO shadercleanup: LazyShader is really a lazyprogram....
 function gl_convert(lazyshader::AbstractLazyShader, data)
     kw_dict = lazyshader.kw_args
     paths = lazyshader.paths
     if all(x-> isa(x, Shader), paths)
         fragdatalocation = get(kw_dict, :fragdatalocation, Tuple{Int, String}[])
-        return compile_program([paths...], fragdatalocation)
+        return Program([paths...], fragdatalocation)
     end
     v = get_view(kw_dict)
     fragdatalocation = get(kw_dict, :fragdatalocation, Tuple{Int, String}[])
@@ -239,7 +186,7 @@ function gl_convert(lazyshader::AbstractLazyShader, data)
             src, _ = template2source(source, v, data)
             compile_shader(Vector{UInt8}(src), typ, :from_string)
         end
-        return compile_program([shaders...], fragdatalocation)
+        return Program([shaders...], fragdatalocation)
     end
     if !all(x-> isa(x, String), paths)
         error("Please supply only paths or tuples of (source, typ) for Lazy Shader
@@ -261,7 +208,7 @@ function gl_convert(lazyshader::AbstractLazyShader, data)
             tr = Dict(zip(template_keys[i], replacements[i]))
             shaders[i] = get_shader!(path, tr, v, data)
         end
-        compile_program(shaders, fragdatalocation)
+        Program(shaders, fragdatalocation)
     end
 end
 
@@ -371,3 +318,193 @@ function glsl_version_string()
         error("could not parse GLSL version: $glsl")
     end
 end
+function iscompiled(shader::GLuint)
+    success = GLint[0]
+    glGetShaderiv(shader, GL_COMPILE_STATUS, success)
+    return first(success) == GL_TRUE
+end
+
+abstract type AbstractProgram end
+mutable struct Program <: AbstractProgram
+    id          ::GLuint
+    shaders     ::Vector{Shader}
+    nametype    ::Dict{Symbol, GLenum}
+    uniformloc  ::Dict{Symbol, Tuple}
+    context     ::AbstractContext
+    function Program(shaders::Vector{Shader}, fragdatalocation::Vector{Tuple{Int, String}})
+        # Remove old shaders
+        exists_context()
+        program = glCreateProgram()::GLuint
+        glUseProgram(program)
+        #attach new ones
+        foreach(shaders) do shader
+            glAttachShader(program, shader.id)
+        end
+
+        #Bind frag data
+        for (location, name) in fragdatalocation
+            glBindFragDataLocation(program, location, ascii(name))
+        end
+
+        #link program
+        glLinkProgram(program)
+        if !islinked(program)
+            for shader in shaders
+                write(STDOUT, shader.source)
+                println("---------------------------")
+            end
+            error(
+                "program $program not linked. Error in: \n",
+                join(map(x-> string(x.name), shaders), " or "), "\n", getinfolog(program)
+            )
+        end
+
+        # generate the link locations
+        nametypedict = uniform_name_type(program)
+        uniformlocationdict = uniformlocations(nametypedict, program)
+        new(program, shaders, nametypedict, uniformlocationdict, current_context())
+    end
+end
+
+function Program(sh_string_typ...)
+    shaders = Shader[]
+    for (source, typ) in sh_string_typ
+        push!(shaders, Shader(gensym(), typ, Vector{UInt8}(source)))
+    end
+    Program(shaders, Tuple{Int, String}[])
+end
+
+
+bind(program::Program) = glUseProgram(program.id)
+unbind(program::AbstractProgram) = glUseProgram(0)
+
+mutable struct LazyProgram <: AbstractProgram
+    sources::Vector
+    data::Dict
+    compiled_program::Union{Program, Void}
+end
+LazyProgram(sources...; data...) = LazyProgram(Vector(sources), Dict(data), nothing)
+
+function Program(lazy_program::LazyProgram)
+    fragdatalocation = get(lazy_program.data, :fragdatalocation, Tuple{Int, String}[])
+    shaders = haskey(lazy_program.data, :arguments) ? Shader.(lazy_program.sources, Ref(lazy_program.data[:arguments])) : Shader.()
+    return Program([shaders...], fragdatalocation)
+end
+
+function bind(program::LazyProgram)
+    iscompiled_orcompile!(program)
+    bind(program.compiled_program)
+end
+
+function iscompiled_orcompile!(program::LazyProgram)
+    if program.compiled_program == nothing
+        program.compiled_program = Program(program)
+    end
+end
+
+##########
+# freeing
+
+# OpenGL has the annoying habit of reusing id's when creating a new context
+# We need to make sure to only free the current one
+function free(x::Program)
+    if !is_current_context(x.context)
+        return # don't free from other context
+    end
+    try
+        glDeleteProgram(x.id)
+    catch e
+        free_handle_error(e)
+    end
+    return
+end
+
+function free_handle_error(e)
+    #ignore, since freeing is not needed if context is not available
+    isa(e, ContextNotAvailable) && return
+    rethrow(e)
+end
+
+islinked(program::GLuint) = glGetProgramiv(program, GL_LINK_STATUS) == GL_TRUE
+
+function createprogram()
+    program = glCreateProgram()
+    @assert program > 0 "couldn't create program. Most likely, opengl context is not active"
+    program::GLuint
+end
+
+shadertype(s::Shader) = s.typ
+function shadertype(f::File{format"GLSLShader"})
+    shadertype(file_extension(f))
+end
+function shadertype(ext::AbstractString)
+    ext == ".comp" && return GL_COMPUTE_SHADER
+    ext == ".vert" && return GL_VERTEX_SHADER
+    ext == ".frag" && return GL_FRAGMENT_SHADER
+    ext == ".geom" && return GL_GEOMETRY_SHADER
+    error("$ext not a valid extension for $f")
+end
+
+#Implement File IO interface
+function load(f::File{format"GLSLShader"})
+    fname = filename(f)
+    source = open(readstring, fname)
+    compile_shader(fname, source)
+end
+function save(f::File{format"GLSLShader"}, data::Shader)
+    s = open(f, "w")
+    write(s, data.source)
+    close(s)
+end
+
+function uniformlocations(nametypedict::Dict{Symbol, GLenum}, program)
+    result = Dict{Symbol, Tuple}()
+    texturetarget = -1 # start -1, as texture samplers start at 0
+    for (name, typ) in nametypedict
+        loc = get_uniform_location(program, name)
+        str_name = string(name)
+        if istexturesampler(typ)
+            texturetarget += 1
+            result[name] = (loc, texturetarget)
+        else
+            result[name] = (loc,)
+        end
+    end
+    return result
+end
+
+const _program_cache = Dict{Any, Program}()
+
+#TODO: shadercleanup: is this necessary?
+gl_convert(shader::Program, data) = shader
+
+
+# function compile_program(shaders, fragdatalocation)
+#     # Remove old shaders
+#     program = createprogram()
+#     glUseProgram(program)
+#     #attach new ones
+#     foreach(shaders) do shader
+#         glAttachShader(program, shader.id)
+#     end
+#
+#     #Bind frag data
+#     for (location, name) in fragdatalocation
+#         glBindFragDataLocation(program, location, ascii(name))
+#     end
+#
+#     #link program
+#     glLinkProgram(program)
+#     if !GLAbstraction.islinked(program)
+#         error(
+#             "program $program not linked. Error in: \n",
+#             join(map(x-> string(x.name), shaders), " or "), "\n", getinfolog(program)
+#         )
+#     end
+#     # Can be deleted, as they will still be linked to Program and released after program gets released
+#     #foreach(glDeleteShader, shader_ids)
+#     # generate the link locations
+#     nametypedict = uniform_name_type(program)
+#     uniformlocationdict = uniformlocations(nametypedict, program)
+#     Program(program, shaders, nametypedict, uniformlocationdict)
+# end
