+++ src/glbackend/GLAbstraction/render.jl
@@ -1,4 +1,4 @@
-render(robj::RenderObject) = render(robj.vao)
+render(robj::RenderObject) = render(robj.vertexarray)
 ##############################################################################################
 #  Generic render functions
 #####
