+++ src/glbackend/GLVisualize/types.jl
@@ -253,6 +253,7 @@ export Intensity, GLIntensity
 
 NOT(x) = !x
 
+#TODO shadercleanup: Why are paths both shaders and paths???????????????????????????????????????????????????????????????????????
 struct GLVisualizeShader <: AbstractLazyShader
     paths::Tuple
     kw_args::Dict{Symbol, Any}
